refactor: eliminate duplication across adapters, vault, and Slack command factories
- Extract ChannelQueue and withRetry to adapters/shared.ts — removes 3 identical copies
- Merge createDirectCommandAdapters + createPrivateSessionCommandAdapters into one
  createCommandAdapters(options = {}) in slack/bot.ts
- Use `const cfg = (this.config ??= { vaults: {} })` in vault.ts addEntry and
  ensureImageSandboxEntry to share config init without a separate helper
- Rename normalizePortalBaseUrl → portalBaseUrl in main.ts (shorter, no redundancy)
- Add comment on activeExecutor mutation pattern in agent.ts

Net: -139 lines
